+++ src/components/Svg/BackIcon.tsx
@@ -1,0 +1,8 @@
+/* eslint-disable @stylistic/max-len */
+import React from "react";
+
+export const BackIcon: React.FC = () => {
+    return (
+        "BackIcon"
+    );
+};